Rename imports folder to fixtures
