+++ src/__tests__/errors.test.ts
@@ -1,6 +1,6 @@
 import {instance as globalContainer} from "../dependency-container";
 import {inject, injectable} from "../decorators";
-import {A01} from "./imports/01-test-case-A01-injects-B01";
+import {A01} from "./fixtures/01-test-case-A01-injects-B01";
 import {errorMatch} from "../error-helpers";
 afterEach(() => {
   globalContainer.reset();
