+++ src/__tests__/fixtures/01-test-case-A01-injects-B01.ts
