+++ src/__tests__/fixtures/01-test-case-B01-injects-A01.ts
